services/voice/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
websockets>=14.0
httpx>=0.28.0
pydantic>=2.10.0
edge-tts>=7.0.0
openai-whisper>=20240930
numpy>=1.26.0
soundfile>=0.13.0
python-multipart>=0.0.18
